refactor: callback page를 views 레이어로 분리
