+++ src/views/callback/index.ts
@@ -1,0 +1,1 @@
+export { default as CallbackPage } from './ui/CallbackPage';
